mem_layer/requirements.txt
mem0ai
# pyttsx3
gtts
playsound
flask
flask-cors
